[clangd] [C++20 Modules] Add --debug-modules-builder to not remove built module files on exit
In practice I found the option is very helpful for me to understand what
happens when clangd's C++20 modules support fails. With '--log=verbose',
I can rerun the command by clangd to understand what's going wrong
actually.

The documentation or adding the option to '--help' list can be done
separately.

+++ clang-tools-extra/clangd/ModulesBuilder.cpp
@@ -14,12 +14,20 @@
 #include "clang/Serialization/ASTReader.h"
 #include "clang/Serialization/ModuleCache.h"
 #include "llvm/ADT/ScopeExit.h"
+#include "llvm/Support/CommandLine.h"
+
 #include <queue>
 
 namespace clang {
 namespace clangd {
 
 namespace {
+
+llvm::cl::opt<bool> DebugModulesBuilder(
+    "debug-modules-builder",
+    llvm::cl::desc("Don't remove clangd's built module files for debugging. "
+                   "Remember to remove them later after debugging."),
+    llvm::cl::init(false));
 
 // Create a path to store module files. Generally it should be:
 //
@@ -122,7 +130,7 @@ struct ModuleFile {
   }
 
   ~ModuleFile() {
-    if (!ModuleFilePath.empty())
+    if (!ModuleFilePath.empty() && !DebugModulesBuilder)
       llvm::sys::fs::remove(ModuleFilePath);
   }
